+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Make it possible to disable `Lint/UnneededDisable`. ([@jonas054][])
 * [#1958](https://github.com/bbatsov/rubocop/issues/1958): Show name of `Lint/UnneededDisable` when `-D/--display-cop-names` is given. ([@jonas054][])
 * Do not show `Style/NonNilCheck` offenses as corrected when the source code is not modified. ([@rrosenblum][])
+* Fix auto-correct in `Style/RedundantReturn` when `return` has no arguments. ([@lumeet][])
 
 ## 0.32.0 (06/06/2015)